Correctif du script de mise à jour des ministères

+++ lib/tasks/mise_a_jour_ministeres_2022.rake
@@ -1,5 +1,5 @@
 desc "Mise à jour des ministères suite au changement de gouvernement en 2022"
-task 20220831_mise_a_jour_ministeres: :environment do
+task mise_a_jour_ministeres_2022: :environment do
   puts "Renommage et mise au propre des ministères inchangés"
 
   Group.where(name: "PREMIER MINISTRE").update(name: "Premier ministre")
@@ -23,23 +23,23 @@ task 20220831_mise_a_jour_ministeres: :environment do
 
   ministeres_a_rattacher = [["MINISTÈRE DE L'ACTION ET DES COMPTES PUBLICS", "Ministère de l'Économie, des Finances et de la Souveraineté industrielle et numérique"], 
     ["MINISTÈRE DE LA COHÉSION DES TERRITOIRES ET DES RELATIONS AVEC LES COLLECTIVITÉS TERRITORIALES", "Ministère de la Transition écologique et de la Cohésion des territoires"]]
-  ancien_ministere = Group.find_by(name: )
-  nouveau_ministere = Group.find_by(name: )
 
   ministeres_a_rattacher.each do |ministere_a_rattacher|
-    ancien_ministere = ministere_a_rattacher[0]
-    nouveau_ministere = ministere_a_rattacher[1]
+    ancien_ministere = Group.find_by(name: ministere_a_rattacher[0])
+    nouveau_ministere = Group.find_by(name: ministere_a_rattacher[1])
   
-    InternshipOffer.where(group_id: ancien_ministere.id).update_all(group_id: nouveau_ministere.id)
-    Organisation.where(group_id: ancien_ministere.id).update_all(group_id: nouveau_ministere.id)
-    Users::MinistryStatistician.where(ministry_id: ancien_ministere.id).update_all(ministry_id: nouveau_ministere.id)
-    EmailWhitelists::Ministry.where(group_id: ancien_ministere.id).update_all(group_id: nouveau_ministere.id)
-    
-    ancien_ministere.destroy
+    if ancien_ministere.present?
+      InternshipOffer.where(group_id: ancien_ministere.id).update_all(group_id: nouveau_ministere.id)
+      Organisation.where(group_id: ancien_ministere.id).update_all(group_id: nouveau_ministere.id)
+      Users::MinistryStatistician.where(ministry_id: ancien_ministere.id).update_all(ministry_id: nouveau_ministere.id)
+      EmailWhitelists::Ministry.where(group_id: ancien_ministere.id).update_all(group_id: nouveau_ministere.id)
+      
+      ancien_ministere.destroy
+    end
   end
 
   puts "Nouveaux ministères"
   
-  Group.create(name: "Ministère de la Transition énergétique")
-  Group.create(name: "Ministère de la Transformation et de la Fonction publiques")
+  Group.find_or_create_by(name: "Ministère de la Transition énergétique", is_public: true, is_paqte: false)
+  Group.find_or_create_by(name: "Ministère de la Transformation et de la Fonction publiques", is_public: true, is_paqte: false)
 end
